Started p4. Refactored a lot of p3 into object.h/c so that I can use it in p4.

+++ hw2.h
@@ -2,13 +2,17 @@
  * hw2.h
  * Contains useful typedefs and macros.
  */
+
 #ifndef HW2_H_
 #define HW2_H_
 
 #define BINARY 1
 #define DIM 2
 #define PIXEL_MAX 255
-#define DEG_PER_RAD 180.0/PI
+#define DEG_PER_RAD (180.0f/PI)
+
+#include <stdlib.h>
+#include <stdio.h>
 
 #include "vision_utilities.h"
 
@@ -24,7 +28,7 @@ typedef struct LabelMap {
 /** Holds an Object's second moment */
 typedef struct SecondMoment {
 	int a, b, c;
-	double thetaMin, thetaMax;
+	float thetaMin, thetaMax;
 } SecondMoment;
 
 /** Holds data on Objects found in an image */
